iox-#1173 Move PublisherInterface into separate file
