+++ iceoryx_posh/test/moduletests/test_popo_smart_chunk.cpp
@@ -14,16 +14,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_posh/popo/publisher.hpp"
-
-#include "iceoryx_hoofs/cxx/unique_ptr.hpp"
-#include "iceoryx_posh/internal/popo/smart_chunk.hpp"
-#include "iceoryx_posh/popo/publisher.hpp"
-#include "iceoryx_posh/popo/request.hpp"
-#include "iceoryx_posh/popo/response.hpp"
-#include "iceoryx_posh/popo/sample.hpp"
-#include "iceoryx_posh/testing/mocks/chunk_mock.hpp"
-
 #include "test.hpp"
 #include "test_popo_smart_chunk_common.hpp"
