+++ iceoryx_posh/test/moduletests/test_popo_smart_chunk_common.hpp
@@ -17,11 +17,14 @@
 #ifndef IOX_TEST_POPO_SMART_CHUNK_COMMON_HPP
 #define IOX_TEST_POPO_SMART_CHUNK_COMMON_HPP
 
+#include "iceoryx_hoofs/cxx/unique_ptr.hpp"
+#include "iceoryx_posh/internal/popo/publisher_interface.hpp"
 #include "iceoryx_posh/internal/popo/rpc_interface.hpp"
-#include "iceoryx_posh/popo/publisher.hpp"
+#include "iceoryx_posh/internal/popo/smart_chunk.hpp"
 #include "iceoryx_posh/popo/request.hpp"
 #include "iceoryx_posh/popo/response.hpp"
 #include "iceoryx_posh/popo/sample.hpp"
+#include "iceoryx_posh/testing/mocks/chunk_mock.hpp"
 
 namespace test_smart_chunk_common
 {
